+++ src/LoansCorp/Loans/Domain/Product/Product.php
@@ -14,7 +14,7 @@ class Product
         private ?string $id,
         private ?string $title,
         private ?int $term,
-        private ?int $interest,
+        private ?float $interest,
     ) {
         $this->loans = new ArrayCollection();
     }
@@ -86,17 +86,17 @@ class Product
     }
 
     /**
-     * @return int|null
+     * @return float|null
      */
-    public function getInterest(): ?int
+    public function getInterest(): ?float
     {
         return $this->interest;
     }
 
     /**
-     * @param int|null $interest
+     * @param float|null $interest
      */
-    public function setInterest(?int $interest): void
+    public function setInterest(?float $interest): void
     {
         $this->interest = $interest;
     }
